RP-83
Avoiding complexity when using the median computation. In case the dataset is even size, there is not way we can have the correct result by not doing two passes on the data (one for splitting the data into 2 subset, the other for computing the upper bound for the correct median computation).


git-svn-id: https://svn.tuebingen.mpg.de/renficiaud/robust_pca/branches/RP-83%20bis@325 3149698f-976c-419d-aa71-eec1decff2c2